[TASK]: refactoring
